Handle pseudo props config in theme
Components can define allowed pseudo props for, for example, inner
components. This pseudo prop accept the styled props for the inner
component they represent. For example if we have a Button component
a possible pseudo prop can be __label that allows the user to specify
styled props for the inner label of the component. This pseudo props
can also be used when defining the default props or variants for a
component.

Also add resolvePropsFor method to Theme to resolver component and
sub components props using the component's default props, variant and
passed props.