Add support for state pseudo props
- Add support for pseudo props when resolving component props. When
a component has an active state (disabled for example) the styled props
set for the given state (through that state pseudo prop, __disabled for
example) will be merge to the resolved props for the component.
- Add __pressed and __disabled pseudo props.
- Add opacity styled prop mapping.

NOTE: State pseudo props have support for inner component's pseudo props
but it doesn't work the other way around. So if for example you want to
style a Button's label when it's disabled you should define the props
as:

```ts
__disabled: { __label: { color: 'white' } }
```

instead of:

```ts
__label: { __disabled: { color: 'white' } }
```

The second example will not work.